// types.ts
export interface SignUpFormValuesType {
  fullName: string;
  email: string;
  password: string;
  phone: string;
  check: boolean;
}

export interface LoginFormValuesType {
  email: string;
  password: string;
}

export interface EditProfileFormValuesType {
  fullName: string;
  email: string;
  phone: string;
}